requirements.txt
sqlalchemy>=1.4
psycopg2-binary
feedparser
requests
python-dotenv
streamlit
pandas
openai